+++ breeze-data/src/main/kotlin/com/windea/breezeframework/data/serializers/xml/JacksonXmlSerializer.kt
@@ -40,5 +40,5 @@ internal object JacksonXmlSerializer : XmlSerializer {
 }
 
 object JacksonXmlSerializerConfig : XmlSerializerConfig {
-	fun configure(builder: (XmlMapper) -> Unit) = builder(JacksonXmlSerializer.mapper)
+	fun configure(block: (XmlMapper) -> Unit) = block(JacksonXmlSerializer.mapper)
 }