+++ breeze-data/src/main/kotlin/com/windea/breezeframework/data/serializers/yaml/JacksonYamlSerializer.kt
@@ -55,5 +55,5 @@ internal object JacksonYamlSerializer : YamlSerializer {
 }
 
 object JacksonYamlSerializerConfig : YamlSerializerConfig {
-	fun configure(builder: (YAMLMapper) -> Unit) = builder(JacksonYamlSerializer.mapper)
+	fun configure(block: (YAMLMapper) -> Unit) = block(JacksonYamlSerializer.mapper)
 }